split 400+ line code into meaningful chunks, also changed 8 parallel lists into 2 dictionaries
